@@ -8,7 +8,7 @@
   ],
   "license": "MIT",
   "require": {
-    "php": "^7.3|^8.0",
+    "php": "^8.0.0",
     "apility/laravel-fcm": "^1.4",
     "fideloper/proxy": "^4.4",
     "fruitcake/laravel-cors": "^2.0",
