@@ -2,16 +2,15 @@
 
 namespace App\Http\Controllers\API;
 
+use App\Actions\User\Authentication\CreateOrGetSocialUser;
 use App\Actions\User\Authentication\LoginUser;
 use App\Actions\User\Authentication\RegisterUser;
-use App\Actions\User\Firebase\RemoveFcmToken;
-use App\Actions\User\Firebase\SetFcmToken;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\LoginUserRequest;
 use App\Http\Requests\RegisterUserRequest;
-use App\Http\Requests\StoreFcmTokenRequest;
 use App\Http\Resources\UserResource;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
 
 class AuthController extends Controller
 {
@@ -22,9 +21,21 @@ class AuthController extends Controller
 
     public function login(LoginUserRequest $request, LoginUser $loginUser): JsonResponse
     {
-        $user = $loginUser->handle($request->validated());
+        $token = $loginUser->handle($request->validated());
 
-        $token = $user->createToken($request->device_name)->plainTextToken;
+        return response()->json([
+            'token' => $token,
+        ]);
+    }
+
+    /*
+     * Provider finished and is returning response
+    */
+    public function social(Request $request, CreateOrGetSocialUser $createOrGetSocialUser): string
+    {
+        $providerPayload = $request->only(['fname', 'lname', 'email', 'avatar', 'provider_id', 'device_name']);
+
+        $token = $createOrGetSocialUser->handle($providerPayload);
 
         return response()->json([
             'token' => $token,
@@ -33,9 +44,7 @@ class AuthController extends Controller
 
     public function register(RegisterUserRequest $request, RegisterUser $registerUser): JsonResponse
     {
-        $user = $registerUser->handle($request->validated());
-
-        $token = $user->createToken($request->device_name)->plainTextToken;
+        $token = $registerUser->handle($request->validated());
 
         return response()->json([
             'token' => $token,
@@ -48,24 +57,6 @@ class AuthController extends Controller
 
         return response()->json([
             'success' => 'Successfully logged out!',
-        ]);
-    }
-
-    public function setFcmToken(StoreFcmTokenRequest $request, SetFcmToken $setFcmToken): JsonResponse
-    {
-        $setFcmToken->handle($request->validated());
-
-        return response()->json([
-            'success' => 'Successfully set FCM token.',
-        ]);
-    }
-
-    public function removeFcmToken(RemoveFcmToken $removeFcmToken): JsonResponse
-    {
-        $removeFcmToken->handle();
-
-        return response()->json([
-            'success' => 'Successfully removed FCM token.',
         ]);
     }
 }
